@@ -6,12 +6,10 @@ import android.content.Context;
 import android.content.Intent;
 import android.media.RingtoneManager;
 import android.net.Uri;
-import android.support.v4.app.NotificationCompat;
-import android.util.Log;
+import androidx.core.app.NotificationCompat;
 
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
-import com.ktu_mantra.ktu.ktumantra_syllabusquestionpaperandcgpacalculator.AboutUsActivity;
 import com.ktu_mantra.ktu.ktumantra_syllabusquestionpaperandcgpacalculator.NotificationActivity;
 import com.ktu_mantra.ktu.ktumantra_syllabusquestionpaperandcgpacalculator.R;
 
