@@ -1,13 +1,14 @@
 package com.ktu_mantra.ktu.ktumantra_syllabusquestionpaperandcgpacalculator;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.View;
 import android.widget.ListView;
 import android.widget.Toast;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 import com.google.android.gms.ads.AdListener;
 import com.google.android.gms.ads.AdRequest;
@@ -21,7 +22,7 @@ import com.ktu_mantra.ktu.ktumantra_syllabusquestionpaperandcgpacalculator.helpe
 import java.util.ArrayList;
 import java.util.List;
 
-public class CalenderActivity extends AppCompatActivity{
+public class CalenderActivity extends AppCompatActivity {
 
     Toolbar mToolbar;
 
