@@ -2,7 +2,6 @@ package com.ktu_mantra.ktu.ktumantra_syllabusquestionpaperandcgpacalculator.adap
 
 import android.content.Context;
 import android.graphics.Color;
-import android.support.v7.widget.CardView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -10,6 +9,8 @@ import android.widget.BaseAdapter;
 import android.widget.ImageButton;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import androidx.cardview.widget.CardView;
 
 import com.ktu_mantra.ktu.ktumantra_syllabusquestionpaperandcgpacalculator.FavouriteDayActivity;
 import com.ktu_mantra.ktu.ktumantra_syllabusquestionpaperandcgpacalculator.R;
